✨ Added navigation to player (if anime is single episode) ✨

+++ feature/translations/src/main/kotlin/club/anifox/android/feature/translations/component/item/param/TranslationComponentItemParam.kt
@@ -5,7 +5,7 @@ import club.anifox.android.core.uikit.param.GlobalParams
 import club.anifox.android.domain.model.anime.translations.AnimeTranslationsCount
 
 internal data class TranslationComponentItemPreviewParam(
-    val onClick: (Int) -> Unit = { },
+    val onClick: () -> Unit = { },
     val animeTranslationsCount: AnimeTranslationsCount,
 )
